Update model to float

+++ phospho-python/phospho/models.py
@@ -798,7 +798,7 @@ class Clustering(ProjectElementBaseModel):
     nb_clusters: Optional[int] = None
     clusters_ids: List[str]
     status: Optional[Literal["started", "summaries", "completed"]] = None
-    percent_of_completion: Optional[int] = None  # 0-100
+    percent_of_completion: Optional[float] = None  # 0-100
     clusters: Optional[List[Cluster]] = None
     scope: Optional[Literal["messages", "sessions"]] = None
     name: Optional[str] = None
